requirements.txt
duckdb==1.1.0
google-genai==0.3.0
pandas==2.2.3
python-dotenv==1.0.1
chromadb==0.5.21
streamlit==1.42.0
